Still need to get female worm burden outputs to match with those produced by R

+++ CHost.h
@@ -17,8 +17,8 @@ public:
 
 	// Members.
 	double birthDate, deathDate;
-	int femaleWorms, totalWorms;
-	int freeliving;
+	double femaleWorms, totalWorms;
+	double freeliving;
 };
 
 #endif /* CHOST_H_ */
